Revert "Split layout settings"
This reverts commit 63a6ddddbf80eeb5edff5bf057f3794f17789b66.
